pyproject.toml
[tool.poetry]
name = "smit-python"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "smit_python"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.0"
pydantic = "^2.0.3"
tortoise-orm = "^0.19.3"
uvicorn = "^0.23.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
